+++ packages/okta-auth-js/lib/constants.js
@@ -8,5 +8,7 @@ module.exports = {
   'REDIRECT_NONCE_COOKIE_NAME': 'okta-oauth-nonce',
   'TOKEN_STORAGE_NAME': 'okta-token-storage',
   'CACHE_STORAGE_NAME': 'okta-cache-storage',
-  'PKCE_STORAGE_NAME': 'okta-pkce-storage'
+  'PKCE_STORAGE_NAME': 'okta-pkce-storage',
+  'ACCESS_TOKEN_STORAGE_KEY': 'accessToken',
+  'ID_TOKEN_STORAGE_KEY': 'idToken'
 };